Extract dentist sprite
